+++ src/plugins/replication-firestore/index.ts
@@ -1,5 +1,5 @@
 /**
- * this plugin adds the RxCollection.syncCouchDBNew()-function to rxdb
+ * this plugin adds the RxCollection.syncCouchDB()-function to rxdb
  * you can use it to sync collections with a remote CouchDB endpoint.
  */
 import {