+++ test/unit.test.ts
@@ -53,7 +53,7 @@ import './unit/orm.test';
 import './unit/replication-protocol.test';
 import './unit/replication.test';
 import './unit/replication-graphql.test';
-import './unit/replication-couchdb-new.test';
+import './unit/replication-couchdb.test';
 import './unit/replication-websocket.test';
 import './unit/replication-p2p.test';
 import './unit/crdt.test';